added router to main page
