+++ App.tsx
@@ -2,20 +2,15 @@ import React, { useState } from 'react';
 import AppLoading from 'expo-app-loading';
 
 // Apollo
-import { ApolloProvider, ApolloClient, InMemoryCache } from '@apollo/client';
+import { ApolloProvider, ApolloClient, createHttpLink } from '@apollo/client';
+import { setContext } from '@apollo/client/link/context';
 import cache from './src/graphql/cache';
-// Navigation
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
 
 // Theme
 import { ThemeProvider } from 'styled-components/native';
 import theme from './src/global/styles/theme';
 
-// Screens
-import { Main } from './src/views/Main';
-import { Login } from './src/views/Login';
-import { Home } from './src/views/Home';
+
 
 // import { NavigationBar } from './src/views/NavigationBar';
 
@@ -26,9 +21,26 @@ import {
   OpenSans_700Bold,
 } from '@expo-google-fonts/open-sans'
 import { Lobster_400Regular } from '@expo-google-fonts/lobster';
+import { Router } from './src/Router';
+
+const httpLink = createHttpLink({
+  uri: 'http://localhost:3000/graphql',
+});
+
+const authLink = setContext((_, { headers }) => {
+  // get the authentication token from local storage if it exists
+  const token = 'sda'
+  // return the headers to the context so httpLink can read them
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : "",
+    }
+  }
+});
 
 const client = new ApolloClient({
-  uri: 'http://localhost:3000/graphql',
+  link: authLink.concat(httpLink),
   cache,
 })
 
@@ -40,32 +52,19 @@ export default function App() {
     Lobster_400Regular,
   });
 
- const [singedIn, setSingedIn] = useState(false);
+  const [singedIn, setSingedIn] = useState(false);
 
-  const Stack = createStackNavigator();
 
+
+  
   if(!fontsLoaded){
     return <AppLoading />
   }
   return  (
     <ThemeProvider theme={theme}>
       <ApolloProvider client={client}>
-      <NavigationContainer>
-        <Stack.Navigator>
-          <>
-            {!singedIn ? (
-              <>
-              <Stack.Screen name="Main" component={Main} options={{ headerShown: false }} />
-              <Stack.Screen name="Login" component={Login} options={{ headerShown: false }} />
-              </>
-            ) : (
-              <Stack.Screen name="Home" component={Home} options={{ headerShown: false }} />
-            )}
-          </>
-        </Stack.Navigator>
-      </NavigationContainer>
+        <Router />
       </ApolloProvider>
-      {/* <NavigationBar /> */}
     </ThemeProvider>
   )
